more navbar icons

+++ app/api/artworks.py
@@ -5,6 +5,7 @@ from app.models import Artwork, db
 from app.forms import ArtworkForm
 from decimal import Decimal
 from datetime import date
+from app.forms import upload_file_to_AWS, get_unique_filename, remove_file_from_AWS
 
 artwork_routes = Blueprint("artworks", __name__)
 
@@ -26,25 +27,51 @@ def single_artwork_route(artwork_id):
     if request.method == 'DELETE':
         if not single_artwork.check_owner(owner_id):
             return {"error": "Forbidden."}, 400
-        db.session.delete(single_artwork)
-        db.session.commit()
-        return {"Success": "Artwork deleted."}, 202
+        file_delete = remove_file_from_AWS(single_artwork.image)
+        if file_delete:
+            db.session.delete(single_artwork)
+            db.session.commit()
+            return {"Success": "Artwork deleted."}, 202
+        else:
+            return {"Error": "Error deleting file image"}, 400
 
     if request.method == "PUT":
         if not single_artwork.check_owner(owner_id):
             return {"error": "Forbidden."}, 400
         form = ArtworkForm()
         form["csrf_token"].data = request.cookies("csrf_token")
+        file_delete = remove_file_from_AWS(single_artwork.image)
+        if not file_delete:
+            return {"Error": "Error deleting file image"}, 400
 
-        if form.validate_on_submit():
-            single_artwork.title = form.data["title"],
-            single_artwork.artist_name = form.data["artist_name"],
-            single_artwork.year = date(int(form.data["year"]), 1, 1),
-            single_artwork.height = Decimal(form.data["height"]),
-            single_artwork.width = Decimal(form.data["width"]),
-            single_artwork.available = True if form.data["available"] == 'True' else False,
-            single_artwork.type = form.data["type"]
-            db.session.commit()
+        if form.data["image"]:
+            new_image = form.data["image"]
+            new_image.filename = get_unique_filename(new_image.filename)
+            upload = upload_file_to_AWS(new_image)
+
+            if "url" not in upload:
+                return {"image": "Image upload failed"}
+
+            if form.validate_on_submit():
+                single_artwork.title = form.data["title"]
+                single_artwork.artist_name = form.data["artist_name"]
+                single_artwork.year = date(int(form.data["year"]), 1, 1)
+                single_artwork.height = Decimal(form.data["height"])
+                single_artwork.width = Decimal(form.data["width"])
+                single_artwork.available = True if form.data["available"] == 'True' else False
+                single_artwork.materials = form.data["materials"]
+                single_artwork.image = upload["url"]
+                db.session.commit()
+        else:
+            if form.validate_on_submit():
+                single_artwork.title = form.data["title"]
+                single_artwork.artist_name = form.data["artist_name"]
+                single_artwork.year = date(int(form.data["year"]), 1, 1)
+                single_artwork.height = Decimal(form.data["height"])
+                single_artwork.width = Decimal(form.data["width"])
+                single_artwork.available = True if form.data["available"] == 'True' else False
+                single_artwork.materials = form.data["materials"]
+                db.session.commit()
 
     return single_artwork.to_safe_dict(), 200
 
@@ -56,6 +83,13 @@ def upload_an_artwork():
     form["csrf_token"].data = request.cookies["csrf_token"]
     owner_id = current_user.id
     if form.validate_on_submit():
+        image = form.data["image"]
+        image.filename = get_unique_filename(image.filename)
+        upload = upload_file_to_AWS(image)
+
+        if "url" not in upload:
+            return {"image": "Image upload failed"}
+
         new_artwork = Artwork(
             title=form.data["title"],
             artist_name=form.data["artist_name"],
@@ -63,7 +97,8 @@ def upload_an_artwork():
             height=Decimal(form.data["height"]),
             width=Decimal(form.data["width"]),
             available=True if form.data["available"] == 'True' else False,
-            type=form.data["type"],
+            materials=form.data["materials"],
+            image=upload["url"],
             owner_id=owner_id
         )
         db.session.add(new_artwork)
